+++ R/CohortComparisonDiagnostics.R
@@ -18,75 +18,55 @@ computeCohortOverlap <- function(connectionDetails = NULL,
                                  connection = NULL,
                                  cohortDatabaseSchema,
                                  cohortTable = "cohort",
-                                 targetCohortId,
-                                 comparatorCohortId) {
-  start <- Sys.time()
+                                 targetCohortIds,
+                                 comparatorCohortIds,
+                                 batchSize = 200) {
+  startTime <- Sys.time()
   
   if (is.null(connection)) {
     connection <- DatabaseConnector::connect(connectionDetails)
     on.exit(DatabaseConnector::disconnect(connection))
   }
   
-  if (!checkIfCohortInstantiated(
-    connection = connection,
-    cohortDatabaseSchema = cohortDatabaseSchema,
-    cohortTable = cohortTable,
-    cohortId = targetCohortId
-  )) {
-    warning(
-      "- Target cohort with ID ",
-      targetCohortId,
-      " appears to be empty. Was it instantiated? Skipping overlap computation."
-    )
-    delta <- Sys.time() - start
-    ParallelLogger::logInfo(paste(
-      "Computing overlap took",
-      signif(delta, 3),
-      attr(delta, "units")
-    ))
-    return(tidyr::tibble())
-  }
+  cohortIds <- c(targetCohortIds, comparatorCohortIds) %>% unique() %>% sort()
   
-  if (!checkIfCohortInstantiated(
-    connection = connection,
-    cohortDatabaseSchema = cohortDatabaseSchema,
-    cohortTable = cohortTable,
-    cohortId = comparatorCohortId
-  )) {
-    warning(
-      "- Comparator cohort with ID ",
-      comparatorCohortId,
-      " appears to be empty. Was it instantiated? Skipping overlap computation."
+  results <- Andromeda::andromeda()
+  for (start in seq(1, length(cohortIds), by = batchSize)) {
+    end <- min(start + batchSize - 1, length(cohortIds))
+    if (length(cohortIds) > batchSize) {
+      ParallelLogger::logInfo(sprintf(
+        "Batch characterization. Processing cohorts %s through %s",
+        start,
+        end
+      ))
+    }
+    
+    sql <- SqlRender::loadRenderTranslateSql(
+      "CohortOverlap.sql",
+      packageName = "CohortDiagnostics",
+      dbms = connection@dbms,
+      cohort_database_schema = cohortDatabaseSchema,
+      cohort_table = cohortTable,
+      target_cohort_ids = targetCohortIds,
+      comparator_cohort_ids = comparatorCohortIds
     )
-    delta <- Sys.time() - start
-    ParallelLogger::logInfo(paste(
-      "Computing overlap took",
-      signif(delta, 3),
-      attr(delta, "units")
-    ))
-    return(tidyr::tibble())
+    overlap <- DatabaseConnector::querySql(
+      connection = connection,
+      sql = sql,
+      snakeCaseToCamelCase = TRUE
+    )
+    if ("overlap" %in% names(results)) {
+      Andromeda::appendToTable(results$overlap, overlap)
+    } else {
+      results$overlap <- overlap
+    }
   }
-  
-  sql <- SqlRender::loadRenderTranslateSql(
-    "CohortOverlap.sql",
-    packageName = "CohortDiagnostics",
-    dbms = connection@dbms,
-    cohort_database_schema = cohortDatabaseSchema,
-    cohort_table = cohortTable,
-    target_cohort_id = targetCohortId,
-    comparator_cohort_id = comparatorCohortId
-  )
-  overlap <- DatabaseConnector::querySql(
-    connection = connection,
-    sql = sql,
-    snakeCaseToCamelCase = TRUE
-  ) %>%
-    tidyr::tibble()
-  delta <- Sys.time() - start
+  overlapAll <- results$overlap %>% dplyr::collect()
+  delta <- Sys.time() - startTime
   ParallelLogger::logInfo(paste(
     "Computing overlap took",
     signif(delta, 3),
     attr(delta, "units")
   ))
-  return(overlap)
+  return(overlapAll)
 }